+++ Src/Core/Shared/Application/Paginations/PaginationDto.cs
@@ -2,12 +2,12 @@ namespace Src.Core.Shared.Application.Paginations;
 
 public record PaginationDto
 {
-    public int Limit { get; }
+    public short Limit { get; }
     public string? StartIndex { get; }
     public string? SortingField { get; }
     public string? SortingType { get; }
 
-    public PaginationDto(int limit, string? startIndex, string? sortingField, string? sortingType)
+    public PaginationDto(short limit, string? startIndex, string? sortingField, string? sortingType)
     {
         Limit = limit;
         StartIndex = startIndex;
